Create log-config.yaml
A ConfigMap is always mounted as readOnly.
The following configuration shows how to mount the log-config ConfigMap onto a Pod called configmap-pod.

The log-config ConfigMap is mounted as a volume, and all contents stored in its log_level entry are mounted into the Pod at path /etc/config/log_level. Note that this path is derived from the volume's mountPath and the path keyed with log_level.